Week 20to22_Project_2/requirements.txt
Flask
Flask-PyMongo
Flet
SpeechRecognition
PyDub
NLTK
SpaCy
transformers
PyMongo
pydub
python-dotenv
requests